# web/www/docs/api-reference/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Saitama API Documentation",
    "description": "Explore resources, tutorials, API docs and dynamic examples to get most out of saitama's developer platform.",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "name": "apiKey",
        "in": "header"
      },
      "authorization": {
        "type": "http",
        "scheme": "Bearer"
      },
      "jwt": {
        "type": "http",
        "scheme": "Bearer"
      }
    },
...
  },
  "paths": {
...
    "/wallets/": {
      "post": {
        "tags": [
          "wallets"
        ],
        "description": "This resource is to create a unique wallet.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "customer": {
                    "anyOf": [
                      {
                        "type": "string",
                        "format": "uuid"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "metadata": {
                    "anyOf": [
                      {
                        "anyOf": [
                          {
                            "type": [
                              "string",
                              "number",
                              "boolean",
                              "null"
                            ]
                          },
                          {
                            "type": "object",
                            "additionalProperties": {}
                          },
                          {
                            "type": "array"
                          }
                        ]
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "address": {
                    "type": "string"
                  },
                  "network": {
                    "type": "string",
                    "format": "uuid"
                  }
                },
                "required": [
                  "network"
                ],
                "additionalProperties": false
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "app": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "customer": {
                          "anyOf": [
                            {
                              "type": "string",
                              "format": "uuid"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "address": {
                          "type": "string"
                        },
                        "generated": {
                          "type": "boolean",
                          "enum": [
                            true
                          ]
                        },
                        "network": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string",
                              "format": "uuid"
                            },
                            "name": {
                              "type": "string",
                              "enum": [
                                "solana",
                                "ethereum",
                                "tron"
                              ]
                            }
                          },
                          "required": [
                            "id",
                            "name"
                          ],
                          "additionalProperties": false
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      },
                      "required": [
                        "id",
                        "app",
                        "customer",
                        "address",
                        "generated",
                        "network",
                        "createdAt",
                        "updatedAt"
                      ],
                      "additionalProperties": false
                    },
                    {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "app": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "customer": {
                          "anyOf": [
                            {
                              "type": "string",
                              "format": "uuid"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "metadata": {
                          "anyOf": [
                            {
                              "anyOf": [
                                {
                                  "type": [
                                    "string",
                                    "number",
                                    "boolean",
                                    "null"
                                  ]
                                },
                                {
                                  "type": "object",
                                  "additionalProperties": {}
                                },
                                {
                                  "type": "array"
                                }
                              ]
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "generated": {
                          "type": "boolean",
                          "enum": [
                            false
                          ]
                        },
                        "network": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "$ref": "#/anyOf/0/properties/network/properties/id"
                            },
                            "name": {
                              "$ref": "#/anyOf/0/properties/network/properties/name"
                            }
                          },
                          "required": [
                            "id",
                            "name"
                          ],
                          "additionalProperties": false
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      },
                      "required": [
                        "id",
                        "app",
                        "customer",
                        "metadata",
                        "generated",
                        "network",
                        "createdAt",
                        "updatedAt"
                      ],
                      "additionalProperties": false
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "wallets"
        ],
        "description": "This resource is to retrieve information about all wallets.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "address",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "query",
            "name": "network",
            "required": false
          },
          {
            "schema": {
              "type": "boolean"
            },
            "in": "query",
            "name": "generated",
            "required": false
          },
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ]
            },
            "in": "query",
            "name": "customer",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "in": "query",
            "name": "createdAt",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "in": "query",
            "name": "updatedAt",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/selectWalletSchema"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}